requirements.txt
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-ollama
langchain-chroma
langchain-google-genai
langchain-google-vertexai
chromadb
pypdf
streamlit